Introduce Pre-Processing integration tests
